+++ draw3d.h
@@ -18,7 +18,7 @@ typedef enum
 
 void SetMaterial(int mat, BOOL force_set);
 void color(Object* obj, BOOL construction, PRESENTATION pres, BOOL locked);
-void color_as(OBJECT obj_type, float color_decay, BOOL construction, PRESENTATION pres, BOOL locked);
+void color_as(OBJECT obj_type, double color_decay, BOOL construction, PRESENTATION pres, BOOL locked);
 void draw_object(Object *obj, PRESENTATION pres, LOCK parent_lock);
 Face *text_face(Text *text, Face *f);
 void CALLBACK Draw(void);
